Operator class (#1895)
* Added an operator class that will provide a simpler approach to
defining the mathematical operation that is executed by a layer.  The
goal is that layers would execute operators that are assigned to them
and that operators act as curried functions (based on parameters
during construction).  Operators should be functional, taking inputs
and producing outputs in the forward direction, and taking error
signalsand inputs to produce new error signals in the backward
direction.

Implemented a data_type and element-wise operator sub-classes that are
specialized on Input/Output data types and tuned for operators that
operate on purely local matrices, respectively.

Rewrote the clamp layer as an operator.

* Added missing CMakeLists.txt

* fix some things; reformat c++

* fix a hidden virtual function warning

* Revert "fix some things; reformat c++"

This reverts commit 5dffd137573e086b3a419e37fafdfa8bd9dd59a9.

* fix things without the pretty formatting

* Updated the operator class to take a vector of input tensors and
produces a vector of output tensors.  This should allow implementation
of operators with different input and output tensor requirements.

* Changed device-specific and type-specific [fp|bp]_compute and
compute_local functions to pass vectors of tensors by value to allow
for efficient type conversion via C++17 magic.

* Added helper functions to build local tensor vectors from distributed
tensor vectors.

* Adding an operator "doer" layer class that will execute operators that
do not interact with weight matrices.  This should replace the
majority of the transform, math, etc.

* Refactored the operator class hierarchy to make the top level operator
object templated on the input and output datatypes.  There should be
no need for a type-erased top level class.  This removes the
data_type_operator and minimizes some of the dynamic dispatch.

* Tweaks to the clang-format

* broken commit: cleanup the operator interface a bit

* cleanup cereal registration for clamp

* CPU stuff compiles

* fix issues related to half

* build all possible targets in bamboo

* fix one more half issue

* split operator layer into decl and impl

* finish factory functions for operators and operator layer.

* untested: update cuda impl of clamp.

* add some debugging checks to the OperatorLayer builder

* add OperatorLayer to the layer factory

* clean some things up with factories; add python for clamp operator

* updates to Python stuff

* add initial bits of testing

* add operator== for matrices

* expand the clamp test to include backprop

* cleanup includes

* add some lifecycle tests for clamp

* clean up matrix helpers

* fix clamp layer test

* start tests for the operator layer

* minor addition of extra sanity assertions

* add backward compatibility to Python Front End

* fix typo error

* fix issue with missing FP16 symbols

* Removed reference to doer.hpp file
Removed doer layer that was replaced by operator_layer.

* remove debugging code

* move clamp layer test to operator layer test; fix a compilation warning

* add cereal registration for OperatorLayer

* remove unused macros

* add some checks to the clamp operator tests

* remove extraneous headers from model.cpp

* remove extra import

* move ProtobufSerializable to its own file

* support for fp16 cereal registration for operators

* cleanup test_case name and tags for operator layer serialization

* Update include/lbann/layers/operator_layer_impl.hpp

Co-authored-by: Thomas R. Benson <benson31@llnl.gov>
Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>