Remove the autogenerated Breathe API documentation. (#1934)
Note that this documentation was (1) completely redundant and (2)
completely inferior to the Doxygen API documentation. The Doxygen API
documentation is still generated, but we now only generate HTML
documentation (the XML was only generated for use by Breathe). This
saves time more than anything else.

I have NOT removed the (fragile) code to hack through the XML
documentation to try to figure out reasonable things to import into
Breathe. Should we decide to restore that (we shouldn't), it's all
still there waiting to be completely rewritten (but again, Doxygen is
the right tool, and the fact that Sphinx looks prettier doesn't change
the fact that it's not useful for C++ documentation).