PROBIESNet (#1981)
* PROBIESNet implementation

* Fixed a bug in the HDF5 data reader where it failed to check to make
sure that the data store was enabled on the command line.  Fixed a bug
where building the packing map would segfault if there were no samples
in the data store.  Fixed a bug where the fetch_data_field did not
check that the number of elements in the requested field matched the
provided Hydrogen matrix height.  Used a helper function for getting
the data store to ensure that there is a valid pointer.

* In the HDF5 data reader special data fetch functions need to use the
preprocessor macros to define the strings for samples, responses, and
labels.  Added accessor functions to allow unit tests to force the
pack function to leave existing data in place.

* Fixed PROBIES model and HDF5 unit tests to use the proper samples and
responses data fields rather than datum and response.  This makes the
behavior of data fields consistent across user defined and predefined
data fields.  Added unit tests for the explicit fetch_datum and
fetch_response function.

* Fixed data field name for responses

* Added tests to check that fetch_label fails

* Updated release notes

Co-authored-by: Brian C. Van Essen <vanessen1@llnl.gov>