Bugfix for the checkpoint lenet unit tests (#2035)
* Updated the checkpoint logic so that when checkpointing during a
validation phase it will use the current epoch and step counts of the
active trainng execution context.  This will allow the validation
checkpoints to be properly ordered between the normal training
checkponts.  Additionally, under the new SGD execution algorithm, the
validation execution context is reset for every validation phase.  As
a result, two helper functions were added to the checkpoint class that
allow the callback to find the current active training execution
context if it was available, otherwise the active execution context is
used.  This ensures that in an inference only run with checkpoint, the
normal behavior should be preserved.

* Cleanup (#21)

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>