Improve build_lbann.sh to enable spack --reuse and Power debug builds (#2062)
* Updated the minimum required Spack version to v0.17.1

Added support for using the --reuse flag for both concretizing and
installing LBANN with Spack.  This should minimize dependency churn.

Changed the explicit concretization to only happen when creating the
environment and installing dependencies.  This avoids superfluous,
back to back concretization events.

Addressed a bug in Spack where if a package is developed, it only
requires the label, not the entire spec.  With the entire speck
multiple, sequential, install commands fail.

Updated the default build for Power architectures at LC to use
clang@12.0.1.  This includs switching over to using ESSL for the BLAS
library to avoid build issues with OpenBLAS on Power.  This is
specifically necessary since trying to build a debug build of LBANN
with GCC is running into a linker error when using the gold linker.
Switching to Clang and lld enables debug builds and debug builds with
address sanitizer.

* In the SMILES data reader test removed use of raw pointers in exchange for unique pointers

* In the SMILES data reader reporting should not be done in the destructor.  Moved to the print statistics function.

* Switched back to using GCC for compiling on Power systems due to runtime segfault