Remove the Cloneable interface from the training algorithms.
There's no practical reason to clone a training algorithm. Namely,
*users* will never clone a training algorithm as there's no way to
implement that operation in the front-end. Moreover, it complicates
semantics. Where training algorithms are given unique identifiers,
which is valuable for user-friendly output, how should that be
propagated to the clone? If there are two instances of the algorithm
with the same name, it is no longer unique.