Formal grouping and testing of protobuf utilities.
Currently, some functions are exposed as templates. This is an
implementation convenience -- the number of valid types is small and
bounded, plus "enums" and "messages". It might be good to just provide
an explicit overload set, plus specialized templates for enums and
message types. Mostly, since the templates are only valid for Protobuf
types (e.g., google::protobuf::int32, which may or may not be
std::int32_t, which may or may not exist (typically it does, but it's
technically optional)), it might be good to make that clear to
users. But I don't think these are "Public API" functions (downstreams
could just as easily write these based on their interpretation of
libprotobuf) so maybe it doesn't matter.