Move data store exchange to primary thread (#2087)
* Moved the data store exchange out of the fetch_data call to allow it
to start and finish in the primary thread.  This should alleviate
issues with MPI_THREAD_MULTIPLE and background I/O.  Furthermore,
there should be no loss in performance since it is being executed in
the same sequence as before.  Additionally, split the data store calls
into two phases that are split between initiating the non-blocking
sends and receives and the subsequent wait statements.  Currently,
these calls are still executed back-to-back to provide equivalent
semantics as the prior implementation.

* Cleanup from rebasing.

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>