Added an interface to the data reader to return a Conduit node (#1987)
* Added an interface to the data reader to return a Conduit node
directly from the data reader to the data coordinator using a new
fetch function interface.  Added a data_packer class that uses the
infrastructure from the HDF5 data reader to extract data fields from
Conduit nodes and pack them into the I/O buffers.  Updated the HDF5
data reader to use the new Conduit fetch function interface.  Added a
new field in the data reader class to indicate that concrete data
readers support the new interface.

* Bug fix to make sure that the buffered data coordinator passes a
vector of conduit nodes of mini-batch size.  Corrected the check of
the number of elements returned.  Also commented out the checks to see
if the received node in the data coordinator is contiguous, or a data
pointer.  Given how the data store will return the data at this point
the best that we can aim for is a compact Conduit layout.

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Changed the data_packer class to just be a namespace.

* Removed features from the HDF5 data reader that are now in the
data_packer namespace and replaced by the fetch conduit node interface
in the data reader.

* Added missing CMakeLists.txt file.

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Fix unused variable warning

* Fix an integer comparison warning

* Some cleanup

* Simplify some of the logic in the data_packer code

* Updated ReleaseNotes.txt

* Removed unnecessary include from header file and moved to cpp file

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>
Co-authored-by: Thomas R. Benson <benson31@llnl.gov>