Tidy up protobuf (#2100)
* Formal grouping and testing of protobuf utilities.

Currently, some functions are exposed as templates. This is an
implementation convenience -- the number of valid types is small and
bounded, plus "enums" and "messages". It might be good to just provide
an explicit overload set, plus specialized templates for enums and
message types. Mostly, since the templates are only valid for Protobuf
types (e.g., google::protobuf::int32, which may or may not be
std::int32_t, which may or may not exist (typically it does, but it's
technically optional)), it might be good to make that clear to
users. But I don't think these are "Public API" functions (downstreams
could just as easily write these based on their interpretation of
libprotobuf) so maybe it doesn't matter.

* Use the new protobuf utilities

* Remove legacy protobuf helpers file

* Layer parents, children, and weights are now repeated string fields

Catch tests pass locally. Need to check Python-based tests.

* cleanup a list-as-string in objective function

* No more space-separated strings of numbers in the layers.

This only addresses the C++ side of the code. I need to clean up the PFE yet.

* Fix a type literal issue in conv

* Patch str_list to get tests to pass

* ValueInitializer uses a list of doubles instead of a string

* Cleanup str_list in the unit tests

* Clear str_list out of the PFE and applications

* Cleanup documentation

* Cleanup an awkward check for spaces in weights names

* Fix some more python stuff

* Remove extraneous list2str uses

* More cleaning up in the apps and tests

* Cleanup legacy prototexts that are still being used in tests

* A few more bits of cleanup.

* Fix a python error

* Cleanup last few unit test errors

* Update applications/graph/node2vec/model/random_projection.py

* address questions from review