Bugfix in Spack for compiler specification (#2104)
* Separated out the spack compiler flag from the rest of the variants
and dependent packages.  This ensures that the compiler is propagated
to the top level LBANN package as well as any co-concretized
packages. Pulled out the high speed linker into a conditional block to
make sure it is only set if the compiler is compatible.  Added the
option to override the compiler from the command line.

Added a spack variant of the host architecture to make sure that spack
builds for the intended system and not some other microarchitecture.

* Added support for new spack feature that allows an environment to
specify that the target architecture should be compatible with the
host. This should prevent the error where builds on a system use an
incompatible target architecture family.

* Update scripts/build_lbann.sh

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Added compiler specific module support

* Check that a minimum git hash exists in the Spack repo before even loading spack

* Added an option to avoid the default spack mirrors.  Added an exit code if the right patches to spack aren't found

* Enable C as a primary language for LBANN

* Added a separate GitLab pipeline for Pascal to test different compiler and build script options.

* Fixed some CI script errors

* Fixed more CI script errors

* Added the center compiler to explicit list of spack external packages.

* Pin to a new version of aluminum to avoid a bug at the head of master.

Add support for properly solving the set of spack packages.

Add PyTest in as a top level package to ensure that NumPy can build.

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>
Co-authored-by: Thomas R. Benson <benson31@llnl.gov>