Bugfix ci (#2136)
* Removed Ray as a valid CI target since it has been decomissioned

* Tweaked the MPI catch testing to use LBANN_NNODES env variable

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Reverted unused commit

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Temporary testing

* Added a new flag, -r, to the build script that allows the user to
request to reuse a spack environment or to create one if it doesn't
already exists.  This is halfway between the behavior right now of the
-d flag that will overwrite any similarly named environment and the
lack of either the -r and -d flags which will use a named environment
and fail if it doesn't exist.

* Change the CI pipeline to use a spack environment named after the
branch rather than the job id.  This makes it possible for subseqent
tests of PRs to reuses the same spack environment and should reduce
the CI build times.

* Tweaked the MPI catch testing to use LBANN_NNODES env variable

* Reverted unused commit

* Apply suggestions from code review

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Temporary testing

* Changed the build_script command to not force a clean build.  This
should accelerate reuse of the same CI pipeline for PRs.

* Switch build script for CI to reuse an existing spack environment

* Remove the cleanup code that removes the spack environment

* Removed debugging scene

* Switch CI to using a temporary build directory.

* Reset back to using a clean build.

* Reverted all of the normal testing

* Switched all of the tests to reuse existing spack environments when possible.

* Removed the code that deleted the spack environment that was just built.

* Removed logic from the build script that created symlinks from a set
of user defined build directories to the actual spack build directory.
This deprecates a feature creating a build directory with a cluster
specific naming.

* Fixed a bug where a bad spack configure command was being issued when there are no GPUs.

* Fixed the spack environment name for the CI pipelines.

* Fixed the spack env names to be unique.

* Added a warning on failure for a copy command.

* Disable Corona until CI is switched to using flux

* Updated the minimum spack commit required to enable the package
requires feature that can be used to inform the concretize of hard
constraints.  Added dependencies on pytest for both py-scipy and
py-numpy to address https://github.com/spack/spack/issues/29447.

Cleaned up gitlab CI code that used to cleanup the spack environment.

* Updated the build_script and CI pipeline to use the new Spack variant
+unit_tests to enable catch2 testing support in LBANN.  This removes
the need to use the spack install --test root feature which is broken
in environments (https://github.com/spack/spack/issues/29447).
Removed the scripts --test flag.

* Updated the minimum Spack hash to include the new unit_test variant on
LBANN.

* Pinned a matched set of protobuf and py-protobuf dependencies at a
more modern version.  Removed customizations for Ray (which was
decommisioned).

* Code cleanup

* Added Pip as a dependency for the Catalyst CI build

* Rolled back the version of protobuf being used to 3.10.0.

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>