Bugfix input layer activations (#2164)
* Fixed the input layer so that it would only resize the activation
matrix if it wasn't already setup to be a view of the
data_coordinator's matrix.  This addresses a signficant performance
bug in the data ingestion where the activation matrix was a view into
the data coordinator's internal buffers.  Then on each step the matrix
would be deallocated, reallocated, then set to a view again.  This
also triggered a performance problem in the memory pool allocator in
Hydrogen, which is not designed to cache very large allocations.

Added guards in the top level data_type_layer to make sure that
matrices that are views of other matrices are not resized.

* Added a release note about the performance fix

* Updated the distconv identity test to force it to keep its error
signals to avoid accessing non-existent previous error signals.