Remove references to Aluminum's MPI_CUDA backend
These are vestigial traces of a bygone era. Most of the MPI_CUDA
backend was split out into the HostTransfer backend. LBANN was only
using codepaths that are now in the HostTransfer backend. This removes
some lingering references to the MPI_CUDA backend that are now
erroneous after this split.