Add a command line argument to optionally hang the LBANN app (#2186)
This hang occurs after library initialization but before any
model/training setup. This is superior to the hang callback because it
allows us to debug almost everything that LBANN does, rather than just
things during or after the training loop.

Because this is a property of the driver and not the LBANN library,
the command line argument is only added in the driver and not in
"options.cpp" or any other such place, and its key to the argument
parser map is not made public via the macro idiom in "options.hpp". It
would be an abuse for library code to reference this.