Improve HDF5 data reader robustness (#2245)
* Added support for ingesting all integer and real format numbers in the
HDF5 data reader.  Switched to using the conduit type conversion
routines for HDF5 to Conduit coercion.

Added warnings if integer formats are not coerced into real formats.

Added a guard to throw an error if character / string based formats
are in the data scheme of an HDF5 sample.

Refactored the HDF5 data reader to separate the load_sample and
load_sample_from_sample_list.  This allows for a simple white box
catch2 test to be added.  Added catch2 tests for loading HRRL HDF5 and
C3FD HDF5 files from "disk".

* Identified an issue in the implementation of the repack / transpose
kernels in both the transforms library and in the HDF5 data reader.
Specifically, the existing implementation assumes a column-major
linearization and the data is actually stored in a row-major
linearization.  Reimplemented the repack function into a proper
row-major approach.  Consolidated the repacking functionality to use
the implementation in the transforms classes.

Added support for explicitly checking that all of the keys and values
in the schema metadata files.  This should explicitly throw a warning
if an unknown key is provided.

Added support for 3D data volumes.

Replaced the 'hwc' keyword for image data with two: 'layout' and
'transpose'.  These fields specifically allow specification of a
channels-first or channels-last encoding, and transposition between a
channels-last to channels-first encoding.

Added a set of tests to test the HDF5 load_sample function.  These
tests will write out a in-memory yaml file to .hdf5 file and then read
that sample from file back into the tests.

Added data reader examples for CosmoFlow, 3D-UNet, and c3fd data sets.

Refactored verification code that compares the fields of two Conduit nodes.

Cleaned up and removed unnecssary test data / schema files.

* Added a test to check that the channels-last to -first transpose works.

* Fixed the count passed to the repack kernels.  It should be number of elements, not number of bytes.

* Apply suggestions from code review

Co-authored-by: Tom Benson <benson31@llnl.gov>

* Apply suggestions from code review

Co-authored-by: Tom Benson <benson31@llnl.gov>

* Add exception header for debug builds

* Fixed the naming of the repack samples to reflect that they are 2D 3channel samples.

* Updated the file ingestion tests for the HDF5 data reader to write
temporary files to a proper temp directory.  Removed spurious data
sample.

* Update src/data_readers/data_reader_HDF5.cpp

Co-authored-by: Tom Benson <benson31@llnl.gov>

---------

Co-authored-by: Tom Benson <benson31@llnl.gov>