Added support to setup the spack load for the aws-ofi plugin for both RCCL and NCCL.
Changed the way that the spack compiler is found when creating the
environment.  Now it is added to the site location, which is the spack
installation.  This allows the spack load command to work without
activating the environment.

Changed the way that the cached files are copied into the build
directory so that they will overwrite existing version of the files.