Disentangle the Caliper annotations from the profiler callback.
In particular, the callback does NOT use Caliper at all. It falls back
on trace-targeted profiling APIs (roctx, nvtx, etc). A warning is
printed if Caliper has been initialized advising users to NOT attempt
to use the profiler callback simultaneously with the "first-class"
Caliper functionality in LBANN. But it's just a warning.