Move some annotations for Caliper around (#2315)
* Move some annotations for Caliper around

* Add caliper annotations to match callback capability

* Add {initialize,finalize}_caliper() to the correct LBANN init/finalize bits

* Clarify Caliper initialization.

* Disentangle the Caliper annotations from the profiler callback.

In particular, the callback does NOT use Caliper at all. It falls back
on trace-targeted profiling APIs (roctx, nvtx, etc). A warning is
printed if Caliper has been initialized advising users to NOT attempt
to use the profiler callback simultaneously with the "first-class"
Caliper functionality in LBANN. But it's just a warning.

* Add output to the LBANN preamble to indicate whether caliper has been initialized.

* Update the Python front-end to accommodate the new Caliper CLI

* Only add Python Caliper options if LBANN has Caliper support

* Fail if caliper options are used without caliper support