Improve spack module detection (#2320)
* Added a new python script that is able to find all of the top level
concretized specs and all of the first level of dependent specs for
LBANN.

Add an external find for git, in case a package really needs to use it
as a dependency.

* Disabling feature for duplicates strategy

* Force Spack to properly reuse dependencies by adding the -f command when concretizing.  Also set the environment to always try to reused dependencies and root packages.

* Documented changes

* Changed the concretize and install commands to only enable the test
options for the root packages, not dependents.  This should fix the
bug where py-pytest has to be installed as a root dependency and thus
requiring extraneous packages that fail to build on some platforms.

Changed the CI scripts to explicitly add the scipy package and then
install pytest via pip.

* Propagate the --clean-build flag to the compiler test.

* Added the correct dependencies flag when the clean build flags are
passed.