Changed the initialization for the data_type_layer's fp_setup_output function so that for layers with parents, they will initialize their output (activation) matrices to be the same width as their parent's output matrices.
For layer's without parents, at setup time the output matrices are set
to the mini_batch_size parameter passed to the setup_data function.
At runtime, they are dynamically checked and will stay at the same
size they are currently at.

Added a function to get the mini-batch size of the parent layer's
outputs.  This function also checks that all parent's have the same
mini-batch size.

Added a function to get the current mini-batch size of the layer's
outputs.