Refactor models to cleanup how mini-batch size is propagated (#2343)
* Changed the initialization for the data_type_layer's fp_setup_output
function so that for layers with parents, they will initialize their
output (activation) matrices to be the same width as their parent's
output matrices.

For layer's without parents, at setup time the output matrices are set
to the mini_batch_size parameter passed to the setup_data function.
At runtime, they are dynamically checked and will stay at the same
size they are currently at.

Added a function to get the mini-batch size of the parent layer's
outputs.  This function also checks that all parent's have the same
mini-batch size.

Added a function to get the current mini-batch size of the layer's
outputs.

* Removed the mini_batch_size parameter from the following functions
in the layer class hierarchy: fp_setup_inputs, fp_setup_outputs, bp_setup_gradient_wrt_inputs
and the distconv_adapter class: fp_setup, bp_setup

* Removed support for calculating a global mini-batch across multiple
models and trainers.  The mini-batch is now strictly contained to a
single trainer.  This deprecates an unused (and old) multi-model
execution mode that predated LTFB.

* Removed the notion of effective mini-batch size versus current
mini-batch size

Remove world master mini-batch adjustment

* Remove model offset field.  No longer necessary since data sets do not
span models.

* Improved error logging

* Added a new current_data_reader field size to the model class.  Use
new field to enure that all nodes in the compute and data flow graph,
solving the sommunication about the last mini-batch size.

* Ensure that the model's current_mini_batch_size field is initialized before setting up the layers.

* mini_batch_size layer no longer needs special handling for fp_setup_outputs.

* Fixed the interplay between the execution algorithm, the input layer,
the model, and the data coordinators to establish the current
mini-batch size.  The SGD execution algorithm will now get the current
mini-batch size from the data readers via the data coordinator.  It
will then cache that value in both the model and the current exectuon
context.  The input layer will now receive the current mini-batch size
directly from the model during fp_setup_data.  Note that the input
layer will no longer be responsible for updating the execution
context.

Corrected the behavior of the batched_inference algorithm so that it
will also explicitly set the current mini-batch size on the model
during execution.

Added documentation in the data ingestion section of the docs to
record the role and ownership of both maximum and current mini-batch
size fields.

* Do not save the m_current_mini_batch_size field in the model checkpoint.

* Update how the distconv adapter gets is mini-batch size.

* Fixed all instances of data coordinator fetch_data() such that the
calling function will set the model and execution context's current
mini-batch size.  This fixes a bug in both check_gradients and kfac.

* Fixed a bug in the interplay between Hydrogen and DistConv
activations. Note that with DistConv adaptors it is possible to have a
parent layer with a zero sized mini-batch.  Once DiHydrogen is the
default, this should be a fatal error.  For the current period, if
there is a layer with a zero sized mini-batch it is probably due to a
distconv layer that skipped setting up the normal activations.  Rely
on the model's current mini-batch size for now.

Added a guard to not setup activations for distconv layers in the
setup_data cleanup phase.

* Removed the duplicative cached value of the current mini-batch size
from the SGD execution context.  Now this field is only cached in the
model and represents the current number of samples that was fetched
from the data reader through the data coordinator.

* In model.hpp changed the type of m_max_mini_batch_size from size_t to
El:Int to match m_current_mini_batch_size.

* Removed the Inter-model communicator callback (imcomm).  This callback
allowed the composition of models with model parallel layers to be
wrapped in an outer-loop data parallelism across multiple models that
would be managed my multiple trainers.  It is superseded by LTFB and
DistConv and will be fully replaced by the DiHydrogen distributed
tensors.

* Update include/lbann/models/model.hpp

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Update src/models/model.cpp

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>