Added support to the data reader to pass the current position in terms of number of steps * mini-batch size into the data readers fetch function hierarchy rather than relying on the internal state of the data reader to determine the current position.  This is a critical fix to enable fetching data in the background, and ahead of time, otherwise the future fetch starts with the current position in the data set.
Note that this is not fully implemented in all of the fetch functions,
specifically the fetch functions that use conduit nodes need to get
this parameter as well.

Updated the data coordinator catch test and it now works with both
synchronous and asynchronous fetch functions.

Note that there seems to be a race condition that occasionally gets
weird results.