Weight Sharding (#2348)
Implements weight and gradient sharding, Also known as FSDP or ZeRO. To shard, specify `sharded = True` in the Weights definition. This PR also makes an internal API change, where `get_values()` of weights and optimizer returns a full view of the matrix (preserving the semantics from before), and `get_values_sharded()` returns the raw, sharded DistMatrix.

Also comments out weight reconciliation as it is not/should not be used in the standard synchronous SGD case.