Double-buffering data coordinator (#2335)
* Make data coordinator double-buffering scheme asynchronous

* Reset minibatch size after receiving input from data coordinator

* Merge fixes and explicit minibatch size

* Add test and do not rely on trainer

* Changed the register_active_data_field so that it no longer calls
setup_data_fields in the data coordinator.  This now has to be
explicitly called after registering new fields, but it improves how
the data coordinator can be tested without an active trainer.  The
input layer will now call the data coordinator's setup_data_fields
function.

Improved the data coordinator's fetch_to_local matrix function to also
setup the size of the current mini-batch. The eliminates a circular
dependency in the fetch_data_in_background function.

* Fixed how the current mini-batch size is retreived from the data
reader and stored in the data coordinator.  Additionally, the
current_mini_batch_size is properly separated from the local
mini-batch size for the current rank.

Corrected the order in which the data reader is updated and the active
buffer index is updates, such that the data reader is updated first
since it requires information from the active buffer.

* Fixed the testing for the last mini-batch size so that if a data
reader was going beyond the number of iterations in the epoch, it
would return a zero size mini-batch.  Added checks to make sure there
were no attempts to fetch data for these cases.

Added tracking to count when a data field of a sample was distributed
into the model.  Temporarily issue a warning if there are unconsumed
fields in the I/O buffer.

Temporarily disable all background data fetching and use only
synchronous fetches.

* Fixed a bug where in a catch test a data reader may try to prime the
data store when it hasn't been setup.

Added a print_config function to the data reader to display standard
configuration parameters.

* Preliminary working version of the data coordinator catch test with a
test data reader.  Works with synchronous data fetch requests.

* Added support to the data reader to pass the current position in terms
of number of steps * mini-batch size into the data readers fetch
function hierarchy rather than relying on the internal state of the
data reader to determine the current position.  This is a critical fix
to enable fetching data in the background, and ahead of time,
otherwise the future fetch starts with the current position in the
data set.

Note that this is not fully implemented in all of the fetch functions,
specifically the fetch functions that use conduit nodes need to get
this parameter as well.

Updated the data coordinator catch test and it now works with both
synchronous and asynchronous fetch functions.

Note that there seems to be a race condition that occasionally gets
weird results.

* Fixed a race condition where the I/O thread fetching the data from the
data reader was using I/O buffer state and data reader state that
could change when the execution algorithm was ready for the next
batch.  If the data reader internal state or data coordinator state
would change, then the background thread may read stale data.  This
has been fixed by caching the size of the current mini-batch, the
current position within the data set, and the reference to the I/O
buffer being used prior to dispatching the thread.

* Fixed the conduit-based data readers fetch interfaces to also take the
current position in the data set as a parameter.

* Fixed the python data reader to use the proper parameter for the
current_position_in_data_set rather than the object's field
m_current_pos to work with background I/O.

* Enable background data fetching in the SGD execution algorithm.

* Removed the num-parallel-readers parameter to the I/O subsystem.
This eliminates an older version of I/O parallelism that relied on
a non-data-parallel I/O buffer and had different ranks fetching
entire mini-batchs.  It is superseded by standard data-parallel I/O.

* Added cleaned up applications and ReleaseNotes.txt for the removal of the num_parallel_readers.

* Updated the CI test python

* Fixed how the data store does the mini-batch exchange.  Now the start
data exchange takes both the relative position from the buffered data
coordinator as well as the current mini-batch size.

* Last of the python files that removed num_parallel_readers.

* Fixed the runtime option to enable and disable background I/O from the
command line.

Added an environment variable to control background I/O.

Set the default number of I/O threads to 4 to avoid overwhelming the
CPU with threads.

Corrected the reporting of the global mini-batch size (no longer local
mini-batch size).

Removed the duplicate fields tracking the state of background I/O from
the model class.

* Fixed a bug in the background I/O where the first mini-batch of each
epoch needs to be synchronously fetched at the start of the epoch,
because the background I/O fetch at the end of the previous epoch will
not fetch into the new epoch.  Additionally, unified the placement of
the synchronous fetch so that it is inside of the caliper calls for
both the training and evaluation scopes.

* Removed data fields and support functions for the loaded mini-batch
size and index, iteration stride, and reset_mini_batch_idx.  These are
no longer used now that the "horizontal" parallelism, where each rank
would fetch a separate mini-batch has been removed.  This has been
replaced with the asynchronous background I/O.

* Removed deprecated fields.

* Changed the I/O and transform pipeline to use a bank of RNGs that
is now indexed by the sample ID in the load sequence, rather than the
I/O thread ID.  This eliminates variablility when using different
numbers of I/O threads.

Added option for LBANN to set the number of OMP threads to modest
default (4) if the environment doesn't specify anything.

* Updated the launchers on GPU accelerated systems to not set the
OMP_NUM_THREADS environment variable.

* Bypass internal default PIP mirrors.

* Improved meta-programming for managing cereal archives.

* Moved state tracking current position in a data set from the data
reader to the dataset class.

* Updated the inference front end.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Added a method to test if a trainer is setup prior to accessing it.
This is particularly useful for catch testing.

Fixed how the random number generation I/O banks are provisioned.

* Improved the upper range on resnet integration test.

* Imported PyTest.

* Renamed fetch_data to fetch_data_asynchronous.

* Map training ranks to the lower half of the generator range and
all other execution modes use the upper half of the bank.
Keep this as one vector to make it easier to serialize and split
it between training and everything else to improve determinism

* Fixed how the split I/O RNG banks are handled.  Created a default path
for testing that access the entire RNG bank when using an invalid
execution mode.

* Improved the accuracy range.

* Added a function to test if a data reader supports background I/O.
Disabled background I/O for the legacy HDF5 data reader.

---------

Co-authored-by: Brian C. Van Essen <vanessen1@llnl.gov>