Cleaning up the way in which random number generators are used between sequential and OMP parallel blocks.
Added explict RNGs for OMP parallel blocks.

Fixed the declaration and initialization of OMP RNGs when compiling in
deterministic mode (which disables OMP parallel blocks).

Refactored how the RNGs are saved to and loaded from checkpoints.

Changed the creation of the I/O RNG banks so that there is one bank
per I/O thread.  Removed the option to explicitly set the number of
I/O RNGs.